requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
marshmallow==3.20.1
pytz==2023.3
gunicorn==21.2.0
werkzeug==2.3.7
SQLAlchemy==2.0.20
email-validator==2.0.0.post2
requests==2.31.0
boto3==1.34.34
python-magic-bin==0.4.14
# OCR Dependencies
easyocr==1.7.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3 